Gateway codes/requirements (1).txt
# requirements.txt
RPi.GPIO
spidev
PyYAML
pyserial
LoRaRF
circuitpython-nrf24l01